pyproject.toml
[tool.poetry]
name = "python-rag"
version = "0.1.0"
description = ""
authors = ["Anand Sandilya <anand.sandilya@quantilope.com>"]
readme = "README.md"
packages = [
    { include = "*", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
transformers = "^4.44.2"
sentence-transformers = "^3.1.1"
faiss-cpu = "^1.8.0.post1"
strawberry-graphql = {extras = ["fastapi"], version = "^0.138.1"}
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pypdf = "^5.0.0"
pytest = "^8.3.3"
langchain-huggingface = "^0.1.0"
langchain-text-splitters = "^0.3.0"
langchain-chroma = "^0.1.4"
langchain-experimental = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-ollama = "^0.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"
flake8 = "^7.1.1"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.poetry.scripts]
dev = "http_api.main:start"
dev_gql = "api.main:start"